requirements.txt
line-bot-sdk
flask
request
numpy